rename models/shelfPreview.ts to models/shelf-preview.ts
