+++ src/models/shelf-preview.ts
