+++ src/reducers/shelf-preview.ts
@@ -1,8 +1,8 @@
 import {fromSpec} from 'compassql/build/src/query/spec';
 import {Action} from '../actions/index';
 import {SHELF_PREVIEW_SPEC, SHELF_PREVIEW_SPEC_DISABLE} from '../actions/shelf-preview';
+import {DEFAULT_SHELF_PREVIEW, ShelfPreview} from '../models/shelf-preview';
 import {fromSpecQuery} from '../models/shelf/spec';
-import {DEFAULT_SHELF_PREVIEW, ShelfPreview} from '../models/shelfPreview';
 
 
 export function shelfPreviewReducer(preview: ShelfPreview = DEFAULT_SHELF_PREVIEW,
